requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
apscheduler==3.10.4
requests==2.31.0
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1